refactor(tests): consolidate roles module tests from 10 to 6 files
Merge governance_context + registry → test_governance.py (37 tests)
Merge resolve_spec_plan_defaults → test_plan.py (10 tests)
Merge run_plan_resolution → test_run.py (12 tests)
Remove empty placeholder TestPlannerSuccessStateChanged

Preserve all 81 passing tests. Fix lint issues:
- Sort imports in test_governance.py
- Break long lines to satisfy ruff line-length rule
